@@ -1,6 +1,6 @@
 {
     "name": "@paytheory/pay-theory-ui",
-    "version": "0.1.258-0",
+    "version": "0.2.33-0",
     "description": "Made with create-react-library",
     "author": "austinzani",
     "license": "MIT",
@@ -47,20 +47,20 @@
         "deploy-release": "npm run post-commit && npm run predeploy-package && npm publish --tag release"
     },
     "devDependencies": {
-        "@babel/core": "^7.16.0",
-        "@babel/preset-env": "^7.16.0",
-        "@babel/preset-react": "^7.16.0",
-        "@testing-library/jest-dom": "^5.15.0",
+        "@babel/core": "^7.17.2",
+        "@babel/preset-env": "^7.16.11",
+        "@babel/preset-react": "^7.16.7",
+        "@testing-library/jest-dom": "^5.16.2",
         "@testing-library/react": "^12.1.2",
         "babel-loader": "^8.2.3",
         "cross-env": "^7.0.3",
-        "eslint": "^8.2.0",
+        "eslint": "^8.8.0",
         "jest-environment-jsdom-sixteen": "^2.0.0",
         "npm-run-all": "^4.1.5",
-        "parcel": "^2.0.1",
-        "react-scripts": "^4.0.3",
-        "webpack": "^5.64.1",
-        "webpack-cli": "^4.9.1",
+        "parcel": "^2.2.1",
+        "react-scripts": "^5.0.0",
+        "webpack": "^5.68.0",
+        "webpack-cli": "^4.9.2",
         "webpack-node-externals": "^3.0.0"
     },
     "files": [
@@ -86,13 +86,13 @@
     },
     "homepage": "https://github.com/pay-theory/pay-theory-ui#readme",
     "peerDependencies": {
-        "jsdom": "^18.1.0",
+        "jsdom": "^19.0.0",
         "moment": "^2.29.1",
         "prop-types": "^15.7.2",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-dropzone": "^11.3.4",
-        "react-router-dom": "^5.2.0",
+        "react-router-dom": "^6.2.1",
         "webfontloader": "^1.6.28"
     }
 }
