@@ -95,7 +95,7 @@ const GlobalStyle = (props) => {
                         background-color: var(--white);
                         display: grid;
                         height: 100%;
-                        grid-template-columns: 290px 1fr;
+                        grid-template-columns: 320px 1fr;
                         overflow: hidden;
                     }
 
