@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-const PayMethodCol = ({ className, row, col, lastFour, brand }) => {
+const PayMethodCol = ({ className, row, col, lastFour, brand, grey }) => {
   const brandClasses = {
     VISA: "pay-theory-card-visa",
     DISCOVER: "pay-theory-card-discover",
@@ -24,7 +24,9 @@ const PayMethodCol = ({ className, row, col, lastFour, brand }) => {
               brandClasses[brand] || "pay-theory-card-default"
             }`}
           />
-          <p>{brand === "CASH" ? "CASH" : `xx${lastFour}`}</p>
+          <p className={grey ? "grey" : ""}>
+            {brand === "CASH" ? "CASH" : `xx${lastFour}`}
+          </p>
         </span>
       </span>
     </td>
