@@ -42,18 +42,33 @@ const openSheet = (identifier) => {
 
 export { closeSheet, openSheet }
 
-const ModalContent = ({
+const Sheet = ({
     identifier,
     children,
     closeAction,
     left,
     top,
-    bottom
+    bottom,
+    alternateClose,
+    header,
+    zPosition
 }) => {
-    const modalForm = identifier ? `${identifier}-sheet-form` : 'sheet-form'
+    const sheetForm = identifier ? `${identifier}-sheet-form` : 'sheet-form'
     const sheet = identifier ? `${identifier}-sheet` : 'sheet'
 
     const position = left ? 'left' : top ? 'top' : bottom ? 'bottom' : 'right'
+    const finalZPosition = zPosition ? 99 + zPosition : 99
+
+    const close = () => {
+        if(alternateClose) {
+            alternateClose()
+        } else {
+            closeSheet(identifier)
+            if (closeAction) {
+                closeAction()
+            }
+        }   
+    }
 
     return (
         <div className='hide-sheet'>
@@ -61,51 +76,44 @@ const ModalContent = ({
                 className='sheet-off'
                 data-testid='sheet-close'
                 id={sheet}
-                onClick={() => {
-                    closeSheet(identifier)
-                    if (closeAction) {
-                        closeAction()
-                    }
-                }}
+                onClick={close}
             />
             <div
                 className={`sheet-form off ${position}`}
                 data-testid='sheet-form'
-                id={modalForm}
+                id={sheetForm}
             >
-                <div className={position} id='sheet-content'>
+                <div className={`sheet-content ${position}`} id='sheet-content'>
                     <div className='sheet-header'>
+                        {left && header ? <h3>{header}</h3> : null}
                         <IconButton
-                            icon='arrow-left'
-                            onClick={() => {
-                                closeSheet(identifier)
-                                if (closeAction) {
-                                    closeAction()
-                                }
-                            }}
+                            icon={left ? 'times' : 'arrow-left'}
+                            onClick={close}
                         />
+                        {!left && header ? <h3>{header}</h3> : null}
                     </div>
                     <div className='sheet-body'>{children}</div>
                 </div>
             </div>
             <style global='true' jsx='true'>
                 {`
-                    #sheet-content {
+                    #${sheetForm} .sheet-content {
                         display: flex;
                         flex-direction: column;
                     }
 
-                    #sheet-content.left,
-                    #sheet-content.right {
+                    #${sheetForm} .sheet-content.left,
+                    #${sheetForm} .sheet-content.right {
                         width: 484px;
+                        height: 100%;
                     }
 
-                    #sheet-content.top,
-                    #sheet-content.bottom {
+                    #${sheetForm} .sheet-content.top,
+                    #${sheetForm} .sheet-content.bottom {
                         height: 484px;
                     }
 
-                    #sheet-content .sheet-header {
+                    #${sheetForm} .sheet-content .sheet-header {
                         width: 100%;
                         display: flex;
                         justify-content: space-between;
@@ -113,24 +121,30 @@ const ModalContent = ({
                         padding: 4px;
                         height: 64px;
                     }
-                    #sheet-content .sheet-header i {
+
+                    #${sheetForm} .sheet-content .sheet-header h3 {
+                        margin: 0px 16px;
+                    }
+
+                    #${sheetForm} .sheet-content .sheet-header i {
                         cursor: pointer;
                         padding: 10px;
                     }
-                    #sheet-content .sheet-body {
+                    #${sheetForm} .sheet-content .sheet-body {
                         margin: 0px 16px 16px;
                         overflow-y: auto;
                         -ms-overflow-style: none;
                         scrollbar-width: none;
+                        height: 100%;
                     }
-                    #sheet-content .sheet-body::-webkit-scrollbar {
+                    #${sheetForm} .sheet-content .sheet-body::-webkit-scrollbar {
                         display: none;
                     }
-                    .sheet-wrapper {
+                    #${sheetForm} .sheet-wrapper {
                         height: 100%;
                         width: 100%;
                     }
-                    .sheet-on {
+                    #${sheet}.sheet-on {
                         display: flex;
                         flex-direction: column;
                         visibility: visible;
@@ -141,24 +155,24 @@ const ModalContent = ({
                         width: 100%;
                         background: var(--black-opaque-32);
                         backdrop-filter: opacity(50%) blur(2px);
-                        z-index: 99;
+                        z-index: ${finalZPosition} ;
                     }
-                    .sheet-off {
+                    #${sheet}.sheet-off {
                         visibility: hidden;
                     }
-                    .sheet-gone {
+                    #${sheet}.sheet-gone {
                         display: none;
                     }
 
                     /* Style for sheet on right */
-                    .sheet-form.on.right {
+                    #${sheetForm}.sheet-form.on.right {
                         top: 0;
                         right: 0;
                         height: 100%;
                         box-shadow: 0px 0px 16px var(--grey);
                     }
 
-                    .sheet-form.off.right {
+                    #${sheetForm}.sheet-form.off.right {
                         top: 0;
                         right: -500px;
                         height: 100%;
@@ -166,14 +180,14 @@ const ModalContent = ({
                     }
 
                     /* Style for sheet on left */
-                    .sheet-form.on.left {
+                    #${sheetForm}.sheet-form.on.left {
                         top: 0;
                         left: 0;
                         height: 100%;
                         box-shadow: 0px 0px 16px var(--grey);
                     }
 
-                    .sheet-form.off.left {
+                    #${sheetForm}.sheet-form.off.left {
                         top: 0;
                         left: -500px;
                         height: 100%;
@@ -181,7 +195,7 @@ const ModalContent = ({
                     }
 
                     /* Style for sheet on top */
-                    .sheet-form.on.top {
+                    #${sheetForm}.sheet-form.on.top {
                         top: 0;
                         right: 0;
                         height: 484px;
@@ -189,7 +203,7 @@ const ModalContent = ({
                         box-shadow: 0px 0px 16px var(--grey);
                     }
 
-                    .sheet-form.off.top {
+                    #${sheetForm}.sheet-form.off.top {
                         top: -500px;
                         right: 0;
                         height: 484px;
@@ -198,7 +212,7 @@ const ModalContent = ({
                     }
 
                     /* Style for sheet on top */
-                    .sheet-form.on.bottom {
+                    #${sheetForm}.sheet-form.on.bottom {
                         bottom: 0;
                         right: 0;
                         height: 484px;
@@ -206,7 +220,7 @@ const ModalContent = ({
                         box-shadow: 0px 0px 16px var(--grey);
                     }
 
-                    .sheet-form.off.bottom {
+                    #${sheetForm}.sheet-form.off.bottom {
                         bottom: -500px;
                         right: 0;
                         height: 484px;
@@ -214,27 +228,27 @@ const ModalContent = ({
                         box-shadow: 0px 0px 0px var(--grey);
                     }
 
-                    #sheet-content form {
+                    #${sheetForm} .sheet-content form {
                         display: flex;
                         flex-direction: column;
                         align-content: center;
                     }
 
-                    .sheet-form {
+                    #${sheetForm}.sheet-form {
                         display: flex;
                         flex-direction: column;
                         visibility: visible;
                         position: fixed;
-                        background-color: var(--grey-2);
+                        background-color: var(--grey-3);
                         overflow-y: auto;
                         overflow-x: hidden;
                         -webkit-transition: box-shadow 0.3s
                             cubic-bezier(0.4, 0, 0.2, 1);
                         transition: all 0.28s cubic-bezier(0.4, 0, 0.2, 1);
-                        z-index: 99;
+                        z-index: ${finalZPosition};
                     }
 
-                    .sheet-form-gone {
+                    #${sheetForm}.sheet-form-gone {
                         display: none;
                     }
 
@@ -248,12 +262,13 @@ const ModalContent = ({
     )
 }
 
-ModalContent.propTypes = {
+Sheet.propTypes = {
     closeAction: PropTypes.func,
     identifier: PropTypes.string.isRequired,
     top: PropTypes.bool,
     bottom: PropTypes.bool,
-    left: PropTypes.bool
+    left: PropTypes.bool,
+    zPosition: PropTypes.number
 }
 
-export default ModalContent
+export default Sheet
